Added a file for abdessamad own functions
